+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- centralize version management (a341420)
 - add loading state to extension import (0348923)
 - add Anthropics API key management (48d96ec)